+++ src/Pipeline/Autoload/DumpAutoload.php
@@ -2,6 +2,7 @@
 
 namespace BrianHenryIE\Strauss\Pipeline\Autoload;
 
+use BrianHenryIE\Strauss\Composer\ComposerPackage;
 use BrianHenryIE\Strauss\Files\File;
 use BrianHenryIE\Strauss\Config\AutoloadConfigInterface;
 use BrianHenryIE\Strauss\Helpers\FileSystem;
@@ -17,8 +18,12 @@ use Composer\Json\JsonFile;
 use Composer\Repository\InstalledFilesystemRepository;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerInterface;
-use Psr\Log\NullLogger;
+use Seld\JsonLint\ParsingException;
 
+/**
+ * @phpstan-import-type AutoloadKeyArray from ComposerPackage
+ * @phpstan-import-type ComposerConfigArray from ComposerPackage
+ */
 class DumpAutoload
 {
     use LoggerAwareTrait;
@@ -26,7 +31,7 @@ class DumpAutoload
     protected AutoloadConfigInterface $config;
 
     protected FileSystem $filesystem;
-    
+
     protected Prefixer $projectReplace;
 
     protected FileEnumerator $fileEnumerator;
@@ -40,10 +45,8 @@ class DumpAutoload
     ) {
         $this->config = $config;
         $this->filesystem = $filesystem;
-        $this->setLogger($logger ?? new NullLogger());
-
+        $this->setLogger($logger);
         $this->projectReplace = $projectReplace;
-
         $this->fileEnumerator = $fileEnumerator;
     }
 
@@ -61,6 +64,8 @@ class DumpAutoload
 
     /**
      * Uses `vendor/composer/installed.json` to output autoload files to `vendor-prefixed/composer`.
+     *
+     * @throws ParsingException
      */
     protected function generatedMainAutoloader(): void
     {
@@ -82,13 +87,25 @@ class DumpAutoload
         Config::$defaultConfig['vendor-dir'] = $relativeTargetDir;
 
         $projectComposerJson = new JsonFile($this->config->getProjectDirectory() . Factory::getComposerFile());
+        /** @var array{require?:array<string,string>,autoload?:AutoloadKeyArray,config?:ComposerConfigArray} $projectComposerJsonArray */
         $projectComposerJsonArray = $projectComposerJson->read();
         if (isset($projectComposerJsonArray['config'], $projectComposerJsonArray['config']['vendor-dir'])) {
             $projectComposerJsonArray['config']['vendor-dir'] = $relativeTargetDir;
         }
 
-        // Do not include the autoload section from the project composer.json in the vendor-prefixed autoloader.
-        if (isset($projectComposerJsonArray['autoload'])) {
+        /**
+         * Loop over all packages that should be included and ensure the root package requires them. Composer only
+         * includes packages in the autoloader that are required by a parent package (including root). Without this
+         * packages that are selectively prefixed are not included in the autoloader.
+         *
+         * @see AutoloadGenerator::filterPackageMap()
+         */
+        foreach ($this->config->getPackagesToPrefix() as $name => $package) {
+            $projectComposerJsonArray['require'][$name] = '*';
+        }
+
+        // Include the project root autoload in the vendor-prefixed autoloader?
+        if (isset($projectComposerJsonArray['autoload']) && !$this->config->isIncludeRootAutoload()) {
             $projectComposerJsonArray['autoload'] = [];
         }
 
@@ -97,9 +114,9 @@ class DumpAutoload
         $package = $composer->getPackage();
 
         /**
-         * Cannot use `$composer->getConfig()`, need to create a new one so the vendor-dir is correct.
+         * Cannot use `$composer->getConfig()`, need to create a new one so the `vendor-dir` is correct.
          */
-        $config = new \Composer\Config(false, $this->config->getProjectDirectory());
+        $config = new Config(false, $this->config->getProjectDirectory());
 
         $config->merge([
             'config' => $projectComposerJsonArray['config'] ?? []
@@ -118,6 +135,7 @@ class DumpAutoload
         $optimize = true; // $input->getOption('optimize') || $config->get('optimize-autoloader');
 
         $installedJsonFile = new JsonFile($this->config->getTargetDirectory() . 'composer/installed.json');
+        /** @var array{dev?:bool} $installedJson */
         $installedJson = $installedJsonFile->read();
         $localRepo = new InstalledFilesystemRepository($installedJsonFile);
 
@@ -128,7 +146,7 @@ class DumpAutoload
         if ($this->config->getVendorDirectory() !== $this->config->getTargetDirectory()) {
             $isDevMode = false;
         } else {
-            $isDevMode = $installedJson['dev'];
+            $isDevMode = (bool) ($installedJson['dev'] ?? false);
         }
         $generator->setDevMode($isDevMode);
 
@@ -215,7 +233,7 @@ class DumpAutoload
             $phpFiles
         );
 
-        $sourceFile = new File(__DIR__);
+        $sourceFile = new File(__DIR__, __DIR__);
         $composerAutoloadNamespaceSymbol = new NamespaceSymbol(
             'Composer\\Autoload',
             $sourceFile