+++ src/Pipeline/Cleanup/InstalledJson.php
@@ -33,7 +33,7 @@ use Seld\JsonLint\ParsingException;
  * @phpstan-type InstalledJsonPackageAuthorArray array{name:string,email:string}
  * @phpstan-type InstalledJsonPackageSupportArray array{issues:string, source:string}
  *
- * @phpstan-type InstalledJsonPackageArray array{name:string, version:string, version_normalized:string, source:InstalledJsonPackageSourceArray, dist:InstalledJsonPackageDistArray, require:array<string,string>, require-dev:array<string,string>, time:string, type:string, installation-source:string, autoload:InstalledJsonPackageAutoloadArray, notification-url:string, license:array<string>, authors:array<InstalledJsonPackageAuthorArray>, description:string, homepage:string, keywords:array<string>, support:InstalledJsonPackageSupportArray, install-path:string}
+ * @phpstan-type InstalledJsonPackageArray array{name:string, version:string, version_normalized:string, source:InstalledJsonPackageSourceArray, dist:InstalledJsonPackageDistArray, require:array<string,string>, require-dev:array<string,string>, time:string, type:string, installation-source:string, autoload?:InstalledJsonPackageAutoloadArray, notification-url:string, license:array<string>, authors:array<InstalledJsonPackageAuthorArray>, description:string, homepage:string, keywords:array<string>, support:InstalledJsonPackageSupportArray, install-path:string}
  *
  * @phpstan-type InstalledJsonArray array{packages:array<InstalledJsonPackageArray>, dev:bool, dev-package-names:array<string>}
  */
@@ -109,6 +109,7 @@ class InstalledJson
     /**
      * @param InstalledJsonArray $installedJsonArray
      * @param array<string,ComposerPackage> $flatDependencyTree
+     * @return InstalledJsonArray
      */
     protected function updatePackagePaths(array $installedJsonArray, array $flatDependencyTree, string $path): array
     {
@@ -129,6 +130,8 @@ class InstalledJson
                 $newInstallPath = $path . str_replace('../', '', $package['install-path']);
 
                 if (!$this->filesystem->directoryExists($newInstallPath)) {
+                    // Should `unset($installedJsonArray['packages'][$key])`?
+                    // Is this post `delete_vendor_packages`?
                     $this->logger->warning('Package directory unexpectedly DOES NOT exist: ' . $newInstallPath);
                     continue;
                 }
@@ -148,6 +151,9 @@ class InstalledJson
 
     /**
      * Remove autoload key entries from `installed.json` whose file or directory does not exist after deleting.
+     *
+     * @param InstalledJsonArray $installedJsonArray
+     * @return InstalledJsonArray
      */
     protected function removeMissingAutoloadKeyPaths(array $installedJsonArray, string $vendorDir, string $installedJsonPath): array
     {
@@ -170,13 +176,12 @@ class InstalledJson
                 $installedJsonArray['packages'][$packageIndex]['autoload'] = [];
             }
             // delete_vendor_files
-            foreach ($installedJsonArray['packages'][$packageIndex]['autoload'] as $type => $autoload) {
-                $pathExistsInPackage = function (string $vendorDir, array $packageArray, string $relativePath) {
-                    return $this->filesystem->exists(
-                        $vendorDir . 'composer/' . $packageArray['install-path'] . '/' . $relativePath
-                    );
-                };
-
+            $pathExistsInPackage = function (string $vendorDir, array $packageArray, string $relativePath) {
+                return $this->filesystem->exists(
+                    $vendorDir . 'composer/' . $packageArray['install-path'] . '/' . $relativePath
+                );
+            };
+            foreach ($installedJsonArray['packages'][$packageIndex]['autoload'] ?? [] as $type => $autoload) {
                 switch ($type) {
                     case 'files':
                     case 'classmap':
@@ -240,12 +245,13 @@ class InstalledJson
      *
      * @param InstalledJsonArray $installedJsonArray
      * @param array<string,ComposerPackage> $flatDependencyTree
+     * @return InstalledJsonArray
      */
     protected function removeMovedPackagesAutoloadKeyFromVendorDirInstalledJson(array $installedJsonArray, array $flatDependencyTree, string $installedJsonPath): array
     {
         /**
          * @var int $key
-         * @var InstalledJsonPackageArray $package
+         * @var InstalledJsonPackageArray $packageArray
          */
         foreach ($installedJsonArray['packages'] as $key => $packageArray) {
             $packageName = $packageArray['name'];
@@ -273,12 +279,13 @@ class InstalledJson
      *
      * @param InstalledJsonArray $installedJsonArray
      * @param array<string,ComposerPackage> $flatDependencyTree
+     * @return InstalledJsonArray
      */
     protected function removeMovedPackagesAutoloadKeyFromTargetDirInstalledJson(array $installedJsonArray, array $flatDependencyTree, string $installedJsonPath): array
     {
         /**
          * @var int $key
-         * @var InstalledJsonPackageArray $package
+         * @var InstalledJsonPackageArray $packageArray
          */
         foreach ($installedJsonArray['packages'] as $key => $packageArray) {
             $packageName = $packageArray['name'];
@@ -311,6 +318,9 @@ class InstalledJson
         return $installedJsonArray;
     }
 
+    /**
+     * @param InstalledJsonArray $installedJsonArray
+     */
     protected function updateNamespaces(array $installedJsonArray, DiscoveredSymbols $discoveredSymbols): array
     {
         $discoveredNamespaces = $discoveredSymbols->getNamespaces();
@@ -453,7 +463,7 @@ class InstalledJson
 
         $this->logger->debug(
             '{installedJsonFilePath} before: {installedJsonArray}',
-            ['installedJsonFilePath' => $installedJsonFile, 'installedJsonArray' => json_encode($installedJsonArray)]
+            ['installedJsonFilePath' => $installedJsonFile->getPath(), 'installedJsonArray' => json_encode($installedJsonArray)]
         );
 
         $installedJsonArray = $this->updatePackagePaths($installedJsonArray, $flatDependencyTree, $this->config->getTargetDirectory());