+++ tests/Issues/StraussIssue65Test.php
@@ -19,6 +19,8 @@ class StraussIssue65Test extends IntegrationTestCase
      */
     public function test_aws_prefixed_functions()
     {
+        $this->markTestIncomplete('found aws/aws-sdk-php[3.268.17] but these were not loaded, because they are affected by security advisories.');
+
         $this->markTestSkippedOnPhpVersionBelow('8.0');
 
         $this->markTestSkippedOnPhpVersionEqualOrAbove('8.2', 'Fatal error: Allowed memory size of 134217728 bytes exhausted');